+++ src/test/java/org/realityforge/replicant/server/ee/rest/ReplicantPollResourceTest.java
@@ -1,11 +1,15 @@
 package org.realityforge.replicant.server.ee.rest;
 
-import java.lang.reflect.Field;
 import java.util.Map;
+import java.util.concurrent.ScheduledFuture;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import javax.enterprise.concurrent.ContextService;
+import javax.enterprise.concurrent.ManagedScheduledExecutorService;
 import javax.ws.rs.container.AsyncResponse;
 import javax.ws.rs.core.Response;
 import org.mockito.ArgumentCaptor;
-import org.realityforge.replicant.server.ee.rest.ReplicantPollResource.SuspendedRequest;
+import org.realityforge.replicant.server.ee.rest.AbstractReplicantPollResource.SuspendedRequest;
 import org.testng.annotations.Test;
 import static org.mockito.Mockito.*;
 import static org.testng.Assert.*;
@@ -16,12 +20,11 @@ public class ReplicantPollResourceTest
   public void poll_dataAvailable()
     throws Exception
   {
-    final ReplicantPollSource source = mock( ReplicantPollSource.class );
-    final ReplicantPollResource resource = newResource( source );
+    final TestReplicantPollResource resource = newResource();
 
     final AsyncResponse response = mock( AsyncResponse.class );
 
-    when( source.poll( "X", 22 ) ).thenReturn( "DATA!" );
+    when( resource._pollSource.poll( "X", 22 ) ).thenReturn( "DATA!" );
 
     resource.poll( response, "X", 22 );
 
@@ -46,12 +49,11 @@ public class ReplicantPollResourceTest
   public void poll_suspendRequest_thenResume()
     throws Exception
   {
-    final ReplicantPollSource source = mock( ReplicantPollSource.class );
-    final ReplicantPollResource resource = newResource( source );
+    final TestReplicantPollResource resource = newResource();
 
     final AsyncResponse response = mock( AsyncResponse.class );
 
-    when( source.poll( "X", 22 ) ).thenReturn( null );
+    when( resource._pollSource.poll( "X", 22 ) ).thenReturn( null );
 
     resource.poll( response, "X", 22 );
 
@@ -66,12 +68,12 @@ public class ReplicantPollResourceTest
     assertEquals( request.getSessionID(), "X" );
     assertEquals( request.getRxSequence(), 22 );
 
-    when( source.poll( "X", 22 ) ).thenReturn( "DATA" );
+    when( resource._pollSource.poll( "X", 22 ) ).thenReturn( "DATA" );
 
     when( response.isSuspended() ).thenReturn( true );
     when( response.isCancelled() ).thenReturn( false );
 
-    getChecker( resource ).run();
+    resource._dataChecker.run();
 
     final ArgumentCaptor<Response> captor = ArgumentCaptor.forClass( Response.class );
     verify( response ).resume( captor.capture() );
@@ -79,7 +81,7 @@ public class ReplicantPollResourceTest
 
     assertEquals( requests.size(), 0 );
 
-    getChecker( resource ).run();
+    resource._dataChecker.run();
     assertEquals( requests.size(), 0 );
   }
 
@@ -87,12 +89,11 @@ public class ReplicantPollResourceTest
   public void poll_suspendRequest_thenException()
     throws Exception
   {
-    final ReplicantPollSource source = mock( ReplicantPollSource.class );
-    final ReplicantPollResource resource = newResource( source );
+    final TestReplicantPollResource resource = newResource();
 
     final AsyncResponse response = mock( AsyncResponse.class );
 
-    when( source.poll( "X", 22 ) ).thenReturn( null );
+    when( resource._pollSource.poll( "X", 22 ) ).thenReturn( null );
 
     resource.poll( response, "X", 22 );
 
@@ -108,19 +109,19 @@ public class ReplicantPollResourceTest
     assertEquals( request.getRxSequence(), 22 );
 
     final Exception exception = new Exception();
-    when( source.poll( "X", 22 ) ).thenThrow( exception );
+    when( resource._pollSource.poll( "X", 22 ) ).thenThrow( exception );
 
     when( response.isSuspended() ).thenReturn( true );
     when( response.isCancelled() ).thenReturn( false );
 
-    getChecker( resource ).run();
+    resource._dataChecker.run();
 
     final ArgumentCaptor<Response> captor = ArgumentCaptor.forClass( Response.class );
     verify( response, times( 1 ) ).resume( captor.capture() );
     assertResponse( exception, captor.getValue() );
     assertEquals( requests.size(), 0 );
 
-    getChecker( resource ).run();
+    resource._dataChecker.run();
     assertEquals( requests.size(), 0 );
   }
 
@@ -128,13 +129,12 @@ public class ReplicantPollResourceTest
   public void poll_exception()
     throws Exception
   {
-    final ReplicantPollSource source = mock( ReplicantPollSource.class );
-    final ReplicantPollResource resource = newResource( source );
+    final TestReplicantPollResource resource = newResource();
 
     final AsyncResponse response = mock( AsyncResponse.class );
 
     final Exception throwable = new Exception();
-    when( source.poll( "X", 22 ) ).thenThrow( throwable );
+    when( resource._pollSource.poll( "X", 22 ) ).thenThrow( throwable );
 
     resource.poll( response, "X", 22 );
 
@@ -143,37 +143,80 @@ public class ReplicantPollResourceTest
     assertResponse( throwable, captor.getValue() );
   }
 
-  protected ReplicantPollResource newResource( final ReplicantPollSource source )
+  @Test
+  public void poll_exceptionWhenSessionComplete()
     throws Exception
   {
-    final ReplicantPollResource resource = new ReplicantPollResource();
-    setField( resource, "_source", source );
-    resource.setupDataChecker();
+    final TestReplicantPollResource resource = newResource();
+
+    final AsyncResponse response = mock( AsyncResponse.class );
+
+    final Exception throwable = new Exception();
+    when( resource._pollSource.poll( "X", 22 ) ).thenThrow( throwable );
+
+    resource._sessionConnected = false;
+    resource.poll( response, "X", 22 );
+
+    final ArgumentCaptor<Response> captor = ArgumentCaptor.forClass( Response.class );
+    verify( response ).resume( captor.capture() );
+    assertResponse( "", captor.getValue() );
+  }
+
+  protected TestReplicantPollResource newResource()
+    throws Exception
+  {
+    final TestReplicantPollResource resource = new TestReplicantPollResource();
+    resource.postConstruct();
     return resource;
   }
 
-  private ReplicantPollResource.PendingDataChecker getChecker( final ReplicantPollResource resource )
-    throws Exception
+  static class TestReplicantPollResource
+    extends AbstractReplicantPollResource
   {
-    return getField( resource, "_dataChecker" );
-  }
+    private final PollSource _pollSource = mock( PollSource.class );
+    PendingDataChecker _dataChecker;
+    boolean _sessionConnected = true;
 
-  private <T> T getField( final ReplicantPollResource resource,
-                          final String fieldName )
-    throws Exception
-  {
-    final Field field = ReplicantPollResource.class.getDeclaredField( fieldName );
-    field.setAccessible( true );
-    return (T) field.get( resource );
-  }
+    @Override
+    ScheduledFuture<?> schedule( final PendingDataChecker dataChecker )
+    {
+      _dataChecker = dataChecker;
+      return mock( ScheduledFuture.class );
+    }
 
-  private void setField( final ReplicantPollResource resource,
-                         final String fieldName,
-                         final Object value )
-    throws Exception
-  {
-    final Field field = ReplicantPollResource.class.getDeclaredField( fieldName );
-    field.setAccessible( true );
-    field.set( resource, value );
+    @Nonnull
+    @Override
+    PollSource createContextualProxyPollSource()
+    {
+      return _pollSource;
+    }
+
+    @Nonnull
+    @Override
+    protected ManagedScheduledExecutorService getScheduledExecutorService()
+    {
+      return mock( ManagedScheduledExecutorService.class );
+    }
+
+    @Nonnull
+    @Override
+    protected ContextService getContextService()
+    {
+      return mock( ContextService.class );
+    }
+
+    @Nullable
+    @Override
+    protected String poll( @Nonnull final String sessionID, final int rxSequence )
+      throws Exception
+    {
+      return _pollSource.poll( sessionID, rxSequence );
+    }
+
+    @Override
+    protected boolean isSessionConnected( @Nonnull final String sessionID )
+    {
+      return _sessionConnected;
+    }
   }
 }